Added another migration file that adds dummy data to database for users, recipes, and recipebooks

+++ migrations/20161015180013-create-recipe-book.js
@@ -17,7 +17,7 @@ module.exports = {
         defaultValue: false
       },
       price:{
-        type: Sequelize.FLOAT,
+        type: Sequelize.DECIMAL(10, 2),
         defaultValue: 0.00
       },
       createdAt: {
